데이터 길이 늘리기 및 정지버튼 구현
데이터 길이 증가 및 전체 모터 긴급정지 버튼 구현
모터 긴급정지는 테스가 필요하다. 아직 테스트를 진행을 안함

+++ freeRtos_uart/testFn.c
@@ -1,9 +1,10 @@
 ﻿
 #include <avr/io.h>
 #include "testFn.h"
+#include "timer.h"
 
 //아두이노 기본 LED 테스트
-void LED_test(uint8_t delay_value){
+void LED_test(uint32_t delay_value){
 	DDRB |= (1 << PB5);
 
 	//LED 켜기 (ON): PORTB의 5번 비트를 1로 설정 (HIGH)
